task(#44): Clean up
- move mapview to it's own component
